+++ Webots/Emil's World/.Weed remover.wbproj
@@ -1,0 +1,9 @@
+Webots Project File version R2022a
+perspectives: 000000ff00000000fd00000002000000010000011c000002fdfc0200000001fb0000001400540065007800740045006400690074006f00720100000015000002fd0000004100ffffff0000000300000780000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000006900ffffff00000662000002fd00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000100000005620100000002010000000100
+sceneTreePerspectives: 000000ff0000000100000002000001fc000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+consoles: Console:All:All
